only delete static file if mount?

+++ lib/jsroutes.rb
@@ -3,26 +3,26 @@ require 'jsroutes/railtie' if defined?(Rails::Railtie)
 
 module JSRoutes
   COPYRIGHT_NOTE = '// JSRoutes, http://github.com/sirlantis/js-routes'
-  
+
   autoload :Rack, 'jsroutes/rack'
-  
+
   mattr_reader :config
-  
+
   @@config = nil
-  
+
   class << self
     delegate :mode, :minify, :path, :global, :to => :config
-    
+
     alias minify? minify
-    
+
     def write?
       mode == :write
     end
-    
+
     def mount?
       mode == :mount
     end
-    
+
     def build
       template_file = File.join(File.dirname(__FILE__), 'templates', 'router.js')
       template = IO.read(template_file)
@@ -36,22 +36,25 @@ module JSRoutes
           file.puts(script)
         end
       end
-    
+
       script
     end
-    
+
     def configure(app)
       @@config = app.config.jsroutes
-      app.config.middleware.use(JSRoutes::Rack) if mount?
-      File.delete(full_output_path) if File.file?(full_output_path)
+
+      if mount?
+        File.delete(full_output_path) if File.file?(full_output_path)
+        app.config.middleware.use(JSRoutes::Rack)
+      end
     end
-    
+
     def mount_url
       "/#{path}"
     end
-  
+
     protected
-  
+
     def converted_routes
       {}.tap do |routes|
         named_routes.each do |name, route|
@@ -62,11 +65,11 @@ module JSRoutes
         end
       end
     end
-  
+
     def named_routes
       Rails.application.routes.named_routes
     end
-  
+
     def full_output_path
       File.join(Rails.public_path, path)
     end
